Add files via upload
This project introduces a real-time image median filter based on FPGA structure. The median filter is a nonlinear image smoothing technology, its main principle is that consider each pixel in the image as the center, build an odd number of samples (3 × 3, 5 × 5, 7 × 7, etc.) observation window around its neighboring area, sort the gray value of each point, and then use the median value instead of the original value.